feat(admin menu): move admin menu according fsd
