+++ src/widgets/admin-menu/ui/AdminMenu.module.scss
@@ -1,0 +1,114 @@
+@mixin item {
+  display: flex;
+  align-items: center;
+  font-size: 13px;
+  margin-bottom: 9px;
+  padding: 9px 15px 9px 10px;
+  border-radius: 6px;
+  border: 1px solid #fff;
+  transition: background-color 0.1s ease;
+}
+
+@mixin active {
+  color: var(--orange-500);
+  background-color: var(--gray-100);
+  border: 1px solid var(--gray-300);
+}
+
+@mixin icon {
+  display: flex;
+  align-items: center;
+  font-size: 17px;
+  width: fit-content;
+  margin-right: 6px;
+}
+
+.menu {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: sticky;
+  top: 95px;
+  margin-top: 20px;
+  margin-right: 32px;
+  .title {
+    font-size: 13.5px;
+    font-weight: 600;
+    padding: 10px 0;
+  }
+  ul {
+    li {
+      position: relative;
+    }
+    .default {
+      @include item;
+      color: #363b37;
+      .icon {
+        @include icon;
+      }
+      &:hover {
+        @include active;
+      }
+    }
+    .active {
+      @include item;
+      @include active;
+      .icon {
+        @include icon;
+      }
+    }
+    .submenu {
+      padding-left: 15px;
+      position: relative;
+      &::after {
+        content: '';
+        position: absolute;
+        top: 0;
+        left: 16px;
+        border: 0.3px solid var(--gray-400);
+        height: 100%;
+      }
+      .submenuDefault {
+        position: relative;
+        display: block;
+        padding: 8px 0px 8px 12px;
+        color: #37363b;
+        font-size: 12px;
+      }
+      .submenuActive {
+        position: relative;
+        display: block;
+        padding: 8px 0px 8px 12px;
+        color: var(--orange-500);
+        font-size: 12px;
+        &:hover {
+          color: var(--orange-500);
+        }
+      }
+      .submenuDefault,
+      .submenuActive {
+        &::after {
+          content: '';
+          top: 50%;
+          left: 2px;
+          position: absolute;
+          height: 1px;
+          width: 8px;
+          transform: translateY(-50%);
+          background-color: var(--gray-400);
+        }
+      }
+    }
+  }
+  .submenuToggle {
+    display: flex;
+    font-size: 12px;
+    position: absolute;
+    right: 7px;
+    cursor: pointer;
+    &:hover {
+      color: #3d3dbf;
+    }
+  }
+}
